Moved Ask search test cases from test_views.py to test_search.py
